Use environment var for the configuration file
Replace the flag mechanism for specifying a configuration file with the
environment variable VIMPCKRC. I find It more convenient for test to
point to a configuration file and better/easier for the user to
configure. This change is reflected in the tests with the use of the
monkeypatch fixture to set environment variable.