Add test for ConfigFile class
I added a unit test for the new ConfigFile class and I also removed the
test_cli.py test file which was a template for future test. But It feels
redundant to make test for click as I'm already testing each function of
the command module, thus I probably won't make them.